Remove download_multiyear_variables function and put that function into download_multiyear.
Change parameter order to match that of ced.download.

Improve documentation.

It looks like the key discrepancy between my function and ced.download is that I am assuming people set either group or download_variables (exclusive or), and he seems to allow users to set both. For now, I'm OK with that. But if the code ever gets merged then this would be an issue.